requirements.txt
flask~=3.0.0
openai~=1.6.1
gevent~=23.9.1
langchain~=0.0.353
python-dotenv~=1.0.0
spacy~=3.7.2
dnspython~=2.4.2
idna~=3.6
httpcore~=1.0.2
httpx~=0.26.0
anyio~=4.2.0
pip~=23.3.1
emoji~=2.9.0
attrs~=23.2.0
distro~=1.9.0
wheel~=0.41.3
Jinja2~=3.1.2
requests~=2.31.0
PyYAML~=6.0.1
multidict~=6.0.4
exceptiongroup~=1.2.0
sniffio~=1.3.0
MarkupSafe~=2.1.3
werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
numpy~=1.24.4
setuptools~=68.2.2
greenlet~=3.0.3
aiohttp~=3.9.1
tenacity~=8.2.3
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
certifi~=2023.11.17
six~=1.16.0
python-dateutil~=2.8.2
tqdm~=4.66.1
urllib3~=2.1.0
pydantic~=2.5.3
chardet~=5.2.0
langsmith~=0.0.77
SQLAlchemy~=2.0.25
jsonpatch~=1.33
marshmallow~=3.20.1
packaging~=23.2
jsonpointer~=2.4
zipp~=3.17.0
beautifulsoup4~=4.12.2
pinecone-client